+++ src/main/kotlin/org/jetbrains/changelog/exceptions/HeaderParseException.kt
@@ -3,8 +3,7 @@
 package org.jetbrains.changelog.exceptions
 
 class HeaderParseException(value: String, unreleasedTerm: String) : Exception(
-    "Header '$value' does not contain version number. " + (
-        "Probably you want set unreleasedTerm to '$value'"
-            .takeIf { value.contains(unreleasedTerm) } ?: ""
-        )
+    "Header '$value' does not contain version number. "
+            + "By default, SemVer format is required (i.e. 1.0.0). To use other formats, like '1.0', adjust the 'changelog.headerParserRegex' property. "
+            + ("Probably you want set unreleasedTerm to '$value'".takeIf { value.contains(unreleasedTerm) } ?: "")
 )